requirements.txt
pytest==7.4.0
pytest-runner==6.0.0
setuptools==68.0.0
twine==4.0.2
wheel==0.40.0
pyyaml==6.0